First push
Ship camera hotkeys changed:
'v' switches to the next camera
'x' switches to the previous camera

Currently working on warping the ship. Right now it only warps to the
camera position on duo when 'w' is pressed. This is not complete, the
ship  should be facing duo.
See the phase 2 instructions for more details.